Better UI

+++ Fib/app/src/main/java/itac/yzu/fib/MainActivity.java
@@ -9,7 +9,7 @@ import android.view.View;
 import org.w3c.dom.Text;
 
 public class MainActivity extends AppCompatActivity {
-    public int i = 0;
+    public long i = 0;
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
@@ -27,14 +27,14 @@ public class MainActivity extends AppCompatActivity {
     public void COMP(View v) {
         setLabel(fib(i), "COMP");
     }
-    public int fib(int f) {
+    public long fib(long f) {
         if (f <= 1) return f;
         else return fib(f-1) + fib(f-2);
     }
     public void RES(View v) {
         setLabel(0, "RES");
     }
-    public void setLabel(int i, String j) {
+    public void setLabel(long i, String j) {
         TextView tv1 = (TextView)findViewById(R.id.iTV);
         TextView tv2 = (TextView)findViewById(R.id.valTV);
         Button btn1 = (Button)findViewById(R.id.incBTN);
@@ -42,15 +42,15 @@ public class MainActivity extends AppCompatActivity {
 
         switch (j) {
             case "I":
-                tv1.setText(Integer.toString(i));
+                tv1.setText(Long.toString(i));
                 break;
             case "COMP":
-                tv2.setText(Integer.toString(i));
+                tv2.setText(Long.toString(i));
                 btn1.setEnabled(false);
                 btn2.setEnabled(false);
                 break;
             case "RES":
-                tv1.setText(Integer.toString(i));
+                tv1.setText(Long.toString(i));
                 tv2.setText("");
                 btn1.setEnabled(true);
                 btn2.setEnabled(true);
